requirements.txt
APScheduler==3.10.3
beautifulsoup4==4.12.2
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
dnspython==2.4.2
docutils==0.20.1
eventlet==0.33.3
falcon==3.1.1
feedparser==6.0.10
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.0.5
Genshi==0.7.7
gevent==23.7.0
greenlet==2.0.2
gunicorn==21.2.0
# html5-parser==0.4.11
html5lib==1.1
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
# loguru==0.7.0
lxml==4.9.3
MarkupSafe==2.1.3
mechanize==0.4.8
paypalrestsdk==1.13.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
rauth==0.7.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
webencodings==0.5.1
Werkzeug==2.3.7
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
